Move shared components from "utils" into "common"
