+++ app/frontend/src/app/ITwinJsApp/ui-framework/UIFramework.tsx
@@ -2,9 +2,9 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
-import { FrameworkVersion, StateManager, ThemeManager } from "@bentley/ui-framework";
 import * as React from "react";
 import { Provider } from "react-redux";
+import { FrameworkVersion, StateManager, ThemeManager } from "@bentley/ui-framework";
 
 export const UIFramework: React.FC = (props) => {
   return (
