+++ app/frontend/src/app/ITwinJsApp/widgets/PropertyGrid.tsx
@@ -12,8 +12,8 @@ import { PropertyCategory, PropertyData, VirtualizedPropertyGridWithDataProvider
 import { Orientation, useDisposable } from "@bentley/ui-core";
 import { Button } from "@itwin/itwinui-react";
 import { appLayoutContext, AppTab } from "../../AppContext";
-import { VerticalStack } from "../../utils/CenteredStack";
-import { AutoSizer } from "../utils/AutoSizer";
+import { VerticalStack } from "../../common/CenteredStack";
+import { AutoSizer } from "../common/AutoSizer";
 
 export interface PropertyGridProps {
   imodel: IModelConnection;
