+++ app/frontend/src/app/common/AsyncActionButton.scss
