+++ app/frontend/src/app/common/CenteredStack.scss
