+++ app/frontend/src/app/common/CenteredStack.tsx
