+++ app/frontend/src/app/common/LoadingIndicator.tsx
